requirements.txt
livekit-agents
livekit-plugins-anthropic
livekit-plugins-elevenlabs
livekit-plugins-silero
openai
python-dotenv
